Convert from Controllers to Minimal APIs
- Replace Controllers folder with single Endpoints.cs
- Use MapGroup for /skills route prefix
- Cleaner Program.cs without AddControllers/MapControllers
